labels FloatTensor

+++ ECAPAModel.py
@@ -27,7 +27,7 @@ class ECAPAModel(nn.Module):
 		lr = self.optim.param_groups[0]['lr']
 		for num, (data, labels) in enumerate(loader, start = 1):
 			self.zero_grad()
-			labels            = torch.LongTensor(labels).cuda()
+			labels            = torch.FloatTensor(labels).cuda()
 			logits = self.model.forward(data.cuda(), aug = True).squeeze()
 			nloss = self.loss(logits, labels)			
 			nloss.backward()
@@ -46,7 +46,7 @@ class ECAPAModel(nn.Module):
 		## Update the learning rate based on the current epcoh
 		loss = 0
 		for num, (data, labels) in enumerate(loader, start = 1):
-			labels            = torch.LongTensor(labels).cuda()
+			labels            = torch.FloatTensor(labels).cuda()
 			logits = self.model.forward(data.cuda(), aug = False).squeeze()
 			nloss = self.loss(logits, labels)			
 			loss += nloss.detach().cpu().numpy()
